+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### New features
+
+- Add `Ethers.get_transaction/2` function to query native chain transaction by transaction hash.
+
 ## v0.2.1 (2024-01-04)
 
 ### New features